Cloud Database Connected...

+++ StudentsManagement.zip_expanded/StudentManagement/src/main/java/StudentManagement/entity/Student.java
@@ -41,7 +41,7 @@ public class Student {
 	private String degree;
 	
 	@Column(name="MOBILENO")
-	private Integer mobileNo;
+	private Long mobileNo;
 	
 	@Column(name="EMAIL")
 	private String email;
